Car model with attached physics
This initial commit adds a car model with a physics setup attached (but cannot be controlled, it is currently still). As a sidenote, the eggs folder is not needed to run the program, but supplies the models currently being or are expected to be used in future in a compressed format, incase they need to be edited before compilation.